requirements.txt
emoji
python-telegram-bot==12.0.0b1